+++ CHANGELOG.md
@@ -37,6 +37,7 @@ https://gist.github.com/ThornDuke/64da76cd4a56b16492d5101691f6108f
   - CUI
 - README.md: badges
 - config/language-configuration.json: indentation after parentheses
+- src/extension.js: commands `Join Lines No Comments` and `Remove Indentation`
 
 ## [1.3.0] - 2025-09-07